+++ crates/block-producer/src/server.rs
@@ -1,16 +1,21 @@
 use std::{collections::HashMap, net::SocketAddr, time::Duration};
 
 use anyhow::{Context, Result};
-use miden_node_proto::generated::{
-    block_producer::api_server,
-    requests::SubmitProvenTransactionRequest,
-    responses::{BlockProducerStatusResponse, SubmitProvenTransactionResponse},
+use miden_node_proto::{
+    generated::{
+        block_producer::api_server,
+        requests::SubmitProvenTransactionRequest,
+        responses::{BlockProducerStatusResponse, SubmitProvenTransactionResponse},
+    },
+    ntx_builder,
 };
 use miden_node_utils::{
     formatting::{format_input_notes, format_output_notes},
-    tracing::grpc::block_producer_trace_fn,
+    tracing::grpc::{OtelInterceptor, block_producer_trace_fn},
 };
-use miden_objects::{transaction::ProvenTransaction, utils::serde::Deserializable};
+use miden_objects::{
+    note::Nullifier, transaction::ProvenTransaction, utils::serde::Deserializable,
+};
 use tokio::{net::TcpListener, sync::Mutex};
 use tokio_stream::wrappers::TcpListenerStream;
 use tonic::Status;
@@ -22,7 +27,7 @@ use crate::{
     COMPONENT, SERVER_MEMPOOL_EXPIRATION_SLACK, SERVER_MEMPOOL_STATE_RETENTION,
     SERVER_NUM_BATCH_BUILDERS,
     batch_builder::BatchBuilder,
-    block_builder::BlockBuilder,
+    block_builder::{BlockBuilder, NtxClient},
     domain::transaction::AuthenticatedTransaction,
     errors::{AddTransactionError, BlockProducerError, StoreError, VerifyTxError},
     mempool::{BatchBudget, BlockBudget, Mempool, SharedMempool},
@@ -40,6 +45,8 @@ pub struct BlockProducer {
     pub block_producer_address: SocketAddr,
     /// The address of the store component.
     pub store_address: SocketAddr,
+    /// The address of the network transaction builder.
+    pub ntx_builder_address: SocketAddr,
     /// The address of the batch prover component.
     pub batch_prover_url: Option<Url>,
     /// The address of the block prover component.
@@ -57,7 +64,6 @@ impl BlockProducer {
     ///       a fatal error is encountered.
     pub async fn serve(self) -> anyhow::Result<()> {
         info!(target: COMPONENT, endpoint=?self.block_producer_address, store=%self.store_address, "Initializing server");
-
         let store = StoreClient::new(self.store_address);
 
         // retry fetching the chain tip from the store until it succeeds.
@@ -93,10 +99,17 @@ impl BlockProducer {
             .await
             .context("failed to bind to block producer address")?;
 
+        let ntx_builder =
+            ntx_builder::Client::connect_lazy(self.ntx_builder_address, OtelInterceptor);
+
         info!(target: COMPONENT, "Server initialized");
 
-        let block_builder =
-            BlockBuilder::new(store.clone(), self.block_prover_url, self.block_interval);
+        let block_builder = BlockBuilder::new(
+            store.clone(),
+            ntx_builder,
+            self.block_prover_url,
+            self.block_interval,
+        );
         let batch_builder = BatchBuilder::new(
             store.clone(),
             SERVER_NUM_BATCH_BUILDERS,
@@ -137,8 +150,13 @@ impl BlockProducer {
                 }
             })
             .id();
+
+        let ntx_builder =
+            ntx_builder::Client::connect_lazy(self.ntx_builder_address, OtelInterceptor);
         let rpc_id = tasks
-            .spawn(async move { BlockProducerRpcServer::new(mempool, store).serve(listener).await })
+            .spawn(async move {
+                BlockProducerRpcServer::new(mempool, store, ntx_builder).serve(listener).await
+            })
             .id();
 
         let task_ids = HashMap::from([
@@ -160,7 +178,6 @@ impl BlockProducer {
 
         // We could abort the other tasks here, but not much point as we're probably crashing the
         // node.
-
         task_result
             .map_err(|source| BlockProducerError::JoinError { task, source })
             .map(|(_, result)| match result {
@@ -182,6 +199,8 @@ struct BlockProducerRpcServer {
     mempool: Mutex<SharedMempool>,
 
     store: StoreClient,
+
+    ntx_builder: NtxClient,
 }
 
 #[tonic::async_trait]
@@ -215,8 +234,12 @@ impl api_server::Api for BlockProducerRpcServer {
 }
 
 impl BlockProducerRpcServer {
-    pub fn new(mempool: SharedMempool, store: StoreClient) -> Self {
-        Self { mempool: Mutex::new(mempool), store }
+    pub fn new(mempool: SharedMempool, store: StoreClient, ntx_client: NtxClient) -> Self {
+        Self {
+            mempool: Mutex::new(mempool),
+            store,
+            ntx_builder: ntx_client,
+        }
     }
 
     async fn serve(self, listener: TcpListener) -> Result<(), tonic::transport::Error> {
@@ -261,11 +284,27 @@ impl BlockProducerRpcServer {
         let inputs = self.store.get_tx_inputs(&tx).await.map_err(VerifyTxError::from)?;
 
         // SAFETY: we assume that the rpc component has verified the transaction proof already.
+        let tx_id = tx.id();
+        let tx_nullifiers: Vec<Nullifier> = tx.nullifiers().collect();
+
         let tx = AuthenticatedTransaction::new(tx, inputs)?;
 
-        self.mempool.lock().await.lock().await.add_transaction(tx).map(|block_height| {
-            SubmitProvenTransactionResponse { block_height: block_height.as_u32() }
-        })
+        // Update the mempool with the new transaction
+        let submit_tx_response =
+            self.mempool.lock().await.lock().await.add_transaction(tx).map(|block_height| {
+                SubmitProvenTransactionResponse { block_height: block_height.as_u32() }
+            });
+
+        let mut ntx_builder = self.ntx_builder.clone();
+        if let Err(err) = ntx_builder.update_network_notes(tx_id, tx_nullifiers.into_iter()).await {
+            error!(
+                target: COMPONENT,
+                message = %err,
+                "error submitting network notes updates to ntx builder"
+            );
+        }
+
+        submit_tx_response
     }
 }
 
@@ -304,10 +343,17 @@ mod test {
         };
         let block_producer_addr = {
             let block_producer_listener =
-                TcpListener::bind("127.0.0.1:0").await.expect("Failed to bind block-producer");
+                TcpListener::bind("127.0.0.1:0").await.expect("failed to bind block-producer");
             block_producer_listener
                 .local_addr()
                 .expect("Failed to get block-producer address")
+        };
+
+        let ntx_builder_address = {
+            let ntx_builder_address = TcpListener::bind("127.0.0.1:0")
+                .await
+                .expect("failed to bind the ntx builder address");
+            ntx_builder_address.local_addr().expect("failed to get ntx builder address")
         };
 
         // start the block producer
@@ -315,6 +361,7 @@ mod test {
             BlockProducer {
                 block_producer_address: block_producer_addr,
                 store_address: store_addr,
+                ntx_builder_address,
                 batch_prover_url: None,
                 block_prover_url: None,
                 batch_interval: Duration::from_millis(500),
@@ -357,7 +404,6 @@ mod test {
             });
             store_runtime
         };
-
         // we need to wait for the exponential backoff of the block producer to connect to the store
         sleep(Duration::from_secs(1)).await;